第一次作业 使用手动创建线程池

+++ src/main/java/geektime/concurrent/race/ThreadRace.java
@@ -56,8 +56,8 @@ public class ThreadRace {
 		
 		SimplePolicy sp = new SimplePolicy();
 		long totalSimple = sp.go();
-		
-		double rate = totalSimple / totalTime;
+
+		double rate = (double) totalSimple / totalTime;
 		System.out.println("自定义A和基准比较: " + rate);
 	}
